api/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
numpy
pillow
python-multipart
faiss-cpu
